move `chia.types.aliases` to `chia.server.aliases` (#19535)
* move `chia.types.aliases` to `chia.server.aliases`

* touchup
