+++ chia/_tests/farmer_harvester/test_farmer_harvester.py
@@ -19,9 +19,9 @@ from chia.plotting.util import PlotsRefreshParameter
 from chia.protocols import farmer_protocol, harvester_protocol
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.rpc.harvester_rpc_client import HarvesterRpcClient
+from chia.server.aliases import FarmerService, HarvesterService
 from chia.server.outbound_message import NodeType, make_msg
 from chia.simulator.block_tools import BlockTools
-from chia.types.aliases import FarmerService, HarvesterService
 from chia.types.peer_info import UnresolvedPeerInfo
 from chia.util.config import load_config
 from chia.util.hash import std_hash
