+++ chia/_tests/fee_estimation/test_fee_estimation_rpc.py
@@ -10,12 +10,12 @@ from chia_rs.sized_ints import uint64
 
 from chia.rpc.full_node_rpc_api import FullNodeRpcApi
 from chia.rpc.full_node_rpc_client import FullNodeRpcClient
+from chia.server.aliases import WalletService
 from chia.simulator.block_tools import BlockTools
 from chia.simulator.full_node_simulator import FullNodeSimulator
 from chia.simulator.simulator_protocol import FarmNewBlockProtocol
 from chia.simulator.start_simulator import SimulatorFullNodeService
 from chia.simulator.wallet_tools import WalletTool
-from chia.types.aliases import WalletService
 from chia.types.blockchain_format.coin import Coin
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
