+++ chia/_tests/plot_sync/test_plot_sync.py
@@ -29,8 +29,8 @@ from chia.plotting.manager import PlotManager
 from chia.plotting.util import add_plot_directory, remove_plot_directory
 from chia.protocols.harvester_protocol import Plot
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
+from chia.server.aliases import FarmerService, HarvesterService
 from chia.simulator.block_tools import BlockTools
-from chia.types.aliases import FarmerService, HarvesterService
 from chia.util.config import create_default_chia_config, lock_and_load_config, save_config
 from chia.util.streamable import _T_Streamable
